Refactor into single module to maybe fix tests
