--- prefect_gcp/cloud_run_job/__init__.py
@@ -1,1 +1,0 @@
-from .gcp_cloud_run_job import CloudRunJob, CloudRunJobResult  # noqa
